chore: (Re)move unused/hardly used functions
