+++ src/gallia/commands/__init__.py
@@ -104,7 +104,7 @@ if sys.platform.startswith("linux"):
     from gallia.commands.discover.find_xcp import CanFindXCP, TcpFindXCP, UdpFindXCP
     from gallia.commands.discover.uds.isotp import IsotpDiscoverer
     from gallia.commands.fuzz.uds.pdu import PDUFuzzer
-    from gallia.commands.primitive.uds.xcp import SimpleTestXCP
+    from gallia.commands.primitive.xcp import SimpleTestXCP
     from gallia.commands.script.vecu import DbVirtualECU, RngVirtualECU
 
     registry += [
@@ -120,8 +120,8 @@ if sys.platform.startswith("linux"):
 
     __all__ += [
         "CanFindXCP",
-        "UDSFindXCP",
-        "TCPFindXCP",
+        "UdpFindXCP",
+        "TcpFindXCP",
         "IsotpDiscoverer",
         "PDUFuzzer",
         "SimpleTestXCP",
